refactor: rename hive_coder to queen and extract queen orchestrator
